Refactor Create singleton class
